requirements.txt
Pillow==9.2.0
Pillow==10.0.0
psutil==5.9.1
selenium==4.10.0
webdriver_manager==3.8.3
